Jubail E-commerce Laravel
Create a simple shopping cart. √
Use the latest version of Laravel as your PHP framework. √
The project should have a registration and login form. √
The user can also choose to login using social media account (Facebook and Gmail). - I apologize I use Github for login social media √
Add a simple product listing (any product will do). √
The product should be saved in the database. Use MySQL as your database. √
The user can purchase the product. - using stripe √
It is up to you if you want to use a template for the design or create your own as long as the webpage is mobile responsive. √
Upload your source code in github https://github.com/ and add a readme instruction. √
Send the link of your project on or before April 10, 2019 √